news_storage/requirements.txt
# Core Database and ORM Dependencies
sqlalchemy[asyncio]==2.0.15
asyncpg==0.27.0
alembic==1.10.4
psycopg2-binary==2.9.6

# Data Validation and Type Handling
pydantic==1.10.7
python-dotenv==1.0.0

# Datetime and Timezone Support
pytz==2023.3

# Optional: For advanced JSON handling
ujson==5.7.0

# Development and Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0

# Type Checking
mypy==1.3.0
sqlalchemy-stubs==0.4

# Recommended: Performance and Debugging
ipython==8.12.0

# Message Queue
aio-pika==9.5.0
